Transformado em funcao.
