+++ exercises/26_3/package.json
@@ -4,7 +4,7 @@
   "description": "Exercicios com fluxo sincrono, entendendo a arquitetura do NodeJS.",
   "main": "index.js",
   "scripts": {
-    "dev": "node index.js",
+    "start": "node index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Rangel Paolo Cardoso Bomfim <paolorangel1996@outlook.com>",
